third_party/blink/web_tests/editing/deleting/5206311-2-expected.txt


The test deletes cells 5 through 9.
Since all the cells in the third row were deleted, the third row is removed.
The cells in the second row are not removed because not all of them are deleted.
The second cell, which used to contain '5', is replaced with the 'world!' text, and the third cell is cleared:
| "
"
| <table>
|   border="1"
|   "
"
|   <tbody>
|     <tr>
|       <td>
|         "1"
|       <td>
|         "2"
|       <td>
|         "3"
|     "
"
|     <tr>
|       <td>
|         "4"
|       <td>
|         class="start"
|         "<#selection-caret>world!"
|       <td>
| "
"


This test deletes cells 5 through 9 of the first table and 1 through 5 of the second.
In the first table, cells 5-6 in the second row are cleared and the last row (cells 7-9) is removed.
In the second table, the first row (cells 1-3) is removed and cells 4-5 in the second row are cleared:
| "
"
| <table>
|   border="1"
|   "
"
|   <tbody>
|     <tr>
|       <td>
|         "1"
|       <td>
|         "2"
|       <td>
|         "3"
|     "
"
|     <tr>
|       <td>
|         "4"
|       <td>
|         class="start"
|         <#selection-caret>
|         <br>
|       <td>
| <table>
|   border="1"
|   <tbody>
|     <tr>
|       <td>
|       <td>
|         class="end"
|       <td>
|         "6"
|     "
"
|     <tr>
|       <td>
|         "7"
|       <td>
|         "8"
|       <td>
|         "9"
|     "
"
| "
"
